% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMgof.R
\name{NMgof}
\alias{NMgof}
\title{search for models and update or overwrite goodness of fit plots}
\usage{
NMgof(
  dir.models,
  dir.diag,
  models,
  fun.find.models,
  fun.repair.data = NULL,
  fun.gof,
  update.only = TRUE,
  secs.rep,
  hours.run,
  script = NULL
)
}
\arguments{
\item{models}{file names to look for in dir.models. Notice, just
file names, not with directories. Use either models or
fun.find.models}

\item{update.only}{Use TRUE to only run on a model if the model
was run since last time GOF plots were generated. You should
use update.only if you have changed the GOF plot function and
want to rerun for all models. You probably don´t want to use
update.only in combination with secs.rep because then you are
asking for a loop of forced updates.}

\item{hours.run}{How many hours would you like to keep looking for
updates every secs.rep seconds? Default is 24 hours if
secs.rep is provided.}

\item{wipe.existing}{Not implemented yet. Intented to mean that
everything existing in the dir will be deleted if NMgof runs
succesfully.}

\item{...}{Arguments sent to fun.gof}
}
\description{
search for models and update or overwrite goodness of fit plots
}
