% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMplotDiags.R
\name{NMplotDiags}
\alias{NMplotDiags}
\title{Plots of obs vs pop predictions and obs vs individual predictions}
\usage{
NMplotDiags(data, by.split = NULL, facet = NULL, colour = NULL)
}
\arguments{
\item{data}{A data set with observations and model predictions}
}
\description{
Plots of obs vs pop predictions and obs vs individual predictions
}
\details{
If EVID is amoung data columns, only EVID==0 will be
considered.
}
