% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMplotFitOld.R
\name{NMplotFitOld}
\alias{NMplotFitOld}
\title{Get one plot of observations, predictions, and means vs time}
\usage{
NMplotFitOld(
  data,
  geom.dv = "p",
  geom.pred = "",
  geom.ipred = "",
  geom.dvmean = "pc",
  geom.predmean = "l",
  geom.ipredmean = "",
  col.grp,
  type.mean = "arithmetic",
  col.id = "ID",
  col.time = "TIME",
  col.ntim = "NOMTIME",
  col.dv = "DV",
  col.ipred = c("IPRED", "IPRE"),
  col.pred = "PRED",
  log.y = F,
  args.plot.mean = NULL,
  filter.rows,
  debug = F
)
}
\arguments{
\item{data}{The dataset to plot.}

\item{geom.dv}{How to plot observations. A string containing any subset of
\itemize{ \item p: points \item l: lines \item c: confidence interval of
mean at each value of col.ntim.  }}

\item{geom.pred}{How to plot population predictions. See geom.dv.}

\item{geom.ipred}{How to plot individual predictions. See geom.dv.}

\item{geom.dvmean}{How to plot means (at each value of col.ntim) of
observations.}

\item{geom.predmean}{How to plot means (at each value of col.ntim) of
population predictions.}

\item{geom.ipredmean}{How to plot means (at each value of col.ntim) of
individual predictions.}

\item{col.grp}{A grouping variable. Means will be grouped by this (and
col.ntim), and plot will be facetted by grp too.}

\item{type.mean}{Passed to pmxtricks::means().}

\item{col.id}{The name of the column containing subject ID's..}

\item{col.time}{name of column to plot observations against.}

\item{col.ntim}{name of column of nominal time used for calculation and
plotting of means.}

\item{col.dv}{name of column containing observations.}

\item{col.ipred}{Name of column containing individual predictions. Default
is searching for IPRED, then IPRE.}

\item{col.pred}{Name of column containing population predictions.}

\item{log.y}{Use logarithmic scale for y-axis?}

\item{args.plot.mean}{List of arguments to be passed to geoms applied to
mean-type data, like mean(PRED) etc.}

\item{filter.rows}{Default is to only plot EVID==0 records. Set to FALSE, if
this filtering is not wantered.}

\item{debug}{debug Start by calling browser()?}
}
\description{
A summary plot of the observations and predictions. These
important diagnostics plots during model development and as to
understand the dynamics of data and model. The function is very flexible
in terms of what combinations of observations, population predictions,
and individual predictions to plot, and how to represent them.
}
\examples{
library(pmxtricks)
library(data.table)
data(pksim1,package="pmxtricks")
pksim2 <- copy(pksim1)
pksim2[,PRED:=DV+rnorm(.N)*10]
pksim2[,NOMTIME:=TIME]
pksim2[,TIME:=round(TIME+rnorm(.N),2)]


pmxtricks:::NMplotFit(pksim2,
           geom.dv = "p",
           geom.pred = "l",
           geom.dvmean = "lc",
           geom.predmean = "p",
           col.ntim = "NOMTIME",
           col.dv = "DV",
           col.ipred = NULL,
           col.pred = "PRED",
           type.mean = "arithmetic",
           debug=FALSE)+
    theme_pp()

}
