% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{writer}
\alias{writer}
\title{One interface for ggwrite and writeFlextab}
\usage{
writer(
  x,
  file,
  save,
  show,
  formats.ft,
  formats.gg,
  script = NULL,
  canvas = "standard",
  time = NULL,
  ...
)
}
\description{
One interface for ggwrite and writeFlextab
}
