% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMgof.R
\name{NMgof}
\alias{NMgof}
\title{search for models and update or overwrite goodness of fit plots}
\usage{
NMgof(
  dir.models,
  dir.diag,
  models,
  fun.find.models,
  fun.repair.data = NULL,
  fun.gof,
  update.only = TRUE,
  secs.rep,
  hours.run,
  formats.ft = "png",
  formats.gg = "pdf",
  script = NULL,
  time.stamp = NULL,
  canvas = "standard",
  args.NMscanData = NULL
)
}
\arguments{
\item{dir.models}{Directory in which to look for models.}

\item{dir.diag}{The directory to write output to.}

\item{models}{file names to look for in dir.models. Notice, just
file names, not with directories. Use either models or
fun.find.models}

\item{fun.find.models}{A function to apply to filenames found in
dir.models before accepting them as output control streams. If
models not specified, the default is to only look at files
ending in .lst.}

\item{fun.gof}{A list of functions to be run. Instead of just a
function, an element in the list can be a new list including
\code{fun} (the function) and \code{args} (arguments with
values to be passed to fun).}

\item{update.only}{Use TRUE to only run on a model if the model
was run since last time GOF plots were generated. You should
use update.only if you have changed the GOF plot function and
want to rerun for all models. You probably don´t want to use
update.only in combination with secs.rep because then you are
asking for a loop of forced updates.}

\item{hours.run}{How many hours would you like to keep looking for
updates every secs.rep seconds? Default is 24 hours if
secs.rep is provided.}

\item{script}{If not NULL, outputs will be stamped with
tracee::createStamp. \code{script} is intended to include a path to
the script from which the code is run.}

\item{time.stamp}{Passed to tracee::writeFlextab and
tracee::ggwrite as the time argument. This is further passed
to tracee::createStamp. It will be the first string in the
sectond line of the stamps. This string will be followed by
the output file name. time.stamp="model" has a special meaning
where the relative path (from working directory) to the output
control stream of the model being processed. In this case,
there will be no timestamp in stamps. The time stamp will also
be suppressed with time.stamp="".}

\item{canvas}{Passed to ggwrite for plots.}

\item{wipe.existing}{Not implemented yet. Intented to mean that
everything existing in the dir will be deleted if NMgof runs
succesfully.}
}
\description{
search for models and update or overwrite goodness of fit plots
}
\details{
functions to be executed by NMgof must take at least one argument: dt. First arg must be the data set to analyze
}
