% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMplotBSV.R
\name{NMplotBSV}
\alias{NMplotBSV}
\title{Generate distribution plots of between-occasion variability terms
from Nonmem}
\usage{
NMplotBSV(
  data,
  regex.eta,
  names.eta = NULL,
  col.id = "ID",
  covs.num,
  covs.char,
  fun.file = identity,
  save = FALSE,
  show = TRUE,
  script = NULL,
  return.data = FALSE,
  title = NULL,
  structure = "flat",
  debug = F
)
}
\arguments{
\item{data}{A dataset - will be converted to data.frame so
data.table is OK.}

\item{regex.eta}{A regular expression defining the naming of the
ETA's of interest.}

\item{col.id}{The name of the id column in data. Default is ID
like Nonmem.}

\item{covs.num}{Names of columns containing numerical covariates
to plot the random effects against.}

\item{covs.char}{Names of columns containing categorical
covariates to plot the random effects against.}

\item{fun.file}{If saving plots, this function can be used to
translate the file names. The inputs given to the function
argument are "iov_pairs.png" and "iov_covs_n.png".}

\item{save}{Save the generated plots?}

\item{script}{If saving the plots, a stamp to add. See ggstamp.}

\item{return.data}{If TRUE, the identified ETA's together with
subject id and covariates will be returned in both wide and
long format. If FALSE, you just get the plots.}

\item{debug}{Start by running browser()?}
}
\description{
Generate distribution plots of between-occasion variability terms
from Nonmem
}
\seealso{
Other Plotting: 
\code{\link{NMplotGOF}()},
\code{\link{NMplotIndProfs}()},
\code{\link{theme_pp_facet}()},
\code{\link{theme_pp}()}
}
\concept{Plotting}
